Moved the custom EventArgs and Exception to their folder structure and removed the ExecuteSecure methods (They will be added back in a future release on a new IRestProcessor implementation)
