+++ src/SimpleRestServices/Core/Exceptions/BadWebRequestException.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace JSIStudios.SimpleRESTServices.Core.Exceptions
+{
+    public class BadWebRequestException : Exception
+    {
+        public BadWebRequestException(string message)
+            : base(message)
+        {
+        }
+
+        public BadWebRequestException(string message, Exception innerException)
+            : base(message, innerException)
+        {
+        }
+    }
+}
