+++ src/SimpleRestServices/Core/Exceptions/HttpResourceNotFoundException.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace JSIStudios.SimpleRESTServices.Core.Exceptions
+{
+    public class HttpResourceNotFoundException : Exception
+    {
+        public HttpResourceNotFoundException(string message)
+            : base(message)
+        {
+        }
+    }
+}
