+++ src/SimpleRestServices/Core/Exceptions/HttpResourceNotModifiedException.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace JSIStudios.SimpleRESTServices.Core.Exceptions
+{
+    public class HttpResourceNotModifiedException : Exception
+    {
+        public HttpResourceNotModifiedException()
+        {
+        }
+    }
+}
