+++ src/SimpleRestServices/Core/IRequestProcessor.cs
@@ -1,7 +1,8 @@
 using System;
 using System.Collections.Specialized;
+using JSIStudios.SimpleRESTServices.Server.EventArgs;
 
-namespace JSIStudios.SimpleRESTServices.Server
+namespace JSIStudios.SimpleRESTServices.Core
 {
     public interface IRequestProcessor
     {
@@ -14,14 +15,5 @@ namespace JSIStudios.SimpleRESTServices.Server
 
         TResult Execute<TResult>(Func<Guid, TResult> callBack);
         TResult Execute<TResult>(Func<Guid, TResult> callBack, NameValueCollection responseHeaders);
-
-        void ExecuteSecure(Action<Guid> callBack);
-        void ExecuteSecure(Action<Guid> callBack, string controllerName);
-        void ExecuteSecure(Action<Guid> callBack, string controllerName, NameValueCollection responseHeaders);
-
-        TResult ExecuteSecure<TResult>(Func<Guid, TResult> callBack);
-        TResult ExecuteSecure<TResult>(Func<Guid, TResult> callBack, string controllerName);
-        TResult ExecuteSecure<TResult>(Func<Guid, TResult> callBack, string controllerName, NameValueCollection responseHeaders);
-
     }
 }
