+++ src/SimpleRestServices/Server/RESTRequestProcessor.cs
@@ -1,10 +1,14 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Diagnostics;
 using System.Net;
 using System.ServiceModel.Web;
 using System.Text;
 using System.Web;
+using JSIStudios.SimpleRESTServices.Core;
+using JSIStudios.SimpleRESTServices.Core.Exceptions;
+using JSIStudios.SimpleRESTServices.Server.EventArgs;
 
 namespace JSIStudios.SimpleRESTServices.Server
 {
@@ -17,60 +21,6 @@ namespace JSIStudios.SimpleRESTServices.Server
 
         #region Interface methods
         
-        public virtual void ExecuteSecure(Action<Guid> callBack) 
-        {
-            ExecuteSecure(callBack, string.Empty);
-        }
-
-        public virtual void ExecuteSecure(Action<Guid> callBack, string controllerName)
-        {
-            ExecuteSecure(callBack, controllerName, null);
-        }
-
-        public virtual void ExecuteSecure(Action<Guid> callBack, string controllerName, NameValueCollection responseHeaders)
-        {
-            try
-            {
-                if (!IsRequestAuthorized())
-                {
-                    ThrowWebFaultException(HttpStatusCode.Unauthorized);
-                }
-            }
-            catch (HttpHeaderNotFoundException ex)
-            {
-                ThrowWebFaultException<string>(string.Format("Http Header Not Found: {0}", ex.Name), HttpStatusCode.BadRequest);
-            }
-
-            Execute(callBack, responseHeaders);
-        }
-
-        public virtual TResult ExecuteSecure<TResult>(Func<Guid, TResult> callBack)
-        {
-            return ExecuteSecure(callBack, string.Empty);
-        }
-
-        public virtual TResult ExecuteSecure<TResult>(Func<Guid, TResult> callBack, string controllerName)
-        {
-            return ExecuteSecure(callBack, controllerName, null);
-        }
-
-        public virtual TResult ExecuteSecure<TResult>(Func<Guid, TResult> callBack, string controllerName, NameValueCollection responseHeaders)
-        {
-            try
-            {
-                if (!IsRequestAuthorized())
-                {
-                    ThrowWebFaultException(HttpStatusCode.Unauthorized);
-                }
-            }
-            catch (HttpHeaderNotFoundException ex)
-            {
-                ThrowWebFaultException<string>(string.Format("Http Header Not Found: {0}", ex.Name), HttpStatusCode.BadRequest);
-            }
-
-            return Execute(callBack, responseHeaders);
-        }
-
         public virtual void Execute(Action<Guid> callBack)
         {
             Execute(callBack, null);
@@ -111,7 +61,7 @@ namespace JSIStudios.SimpleRESTServices.Server
 
         #region Private methods
 
-        private void ExecuteSafely<TResult>(Func<Guid, TResult> callBack, NameValueCollection responseHeaders)
+        protected void ExecuteSafely<TResult>(Func<Guid, TResult> callBack, NameValueCollection responseHeaders)
         {
             var requestId = Guid.NewGuid();
 
@@ -130,6 +80,9 @@ namespace JSIStudios.SimpleRESTServices.Server
                 if (RequestCompleted != null)
                     RequestCompleted(this, new RESTRequestCompletedEventArgs(requestId, result, stopwatch.ElapsedMilliseconds));
 
+                if (EqualityComparer<TResult>.Default.Equals(result, default(TResult)))
+                    throw new HttpResourceNotFoundException("Resource not found.");
+
                 if (WebOperationContext.Current != null)
                 {
                     WebOperationContext.Current.OutgoingResponse.StatusCode = HttpStatusCode.OK;
@@ -142,14 +95,14 @@ namespace JSIStudios.SimpleRESTServices.Server
                 if (OnError != null)
                     OnError(this, new RESTRequestErrorEventArgs(requestId, ex));
 
-                ThrowWebFaultException<string>(ex.Message, HttpStatusCode.BadRequest);
+                SetHttpErrorStatusCode(ex.Message, HttpStatusCode.BadRequest);
             }
             catch (HttpResourceNotFoundException ex)
             {
                 if (OnError != null)
                     OnError(this, new RESTRequestErrorEventArgs(requestId, ex));
 
-                ThrowWebFaultException<string>(ex.Message, HttpStatusCode.NotFound);
+                SetHttpErrorStatusCode(ex.Message, HttpStatusCode.NotFound);
             }
             catch (HttpResourceNotModifiedException ex)
             {
@@ -159,7 +112,7 @@ namespace JSIStudios.SimpleRESTServices.Server
                 if (WebOperationContext.Current != null)
                     WebOperationContext.Current.OutgoingResponse.StatusCode = HttpStatusCode.NotModified;
                 else
-                    ThrowWebFaultException(HttpStatusCode.NotModified);
+                    SetHttpErrorStatusCode(HttpStatusCode.NotModified);
             }
             catch (WebFaultException)
             {
@@ -173,7 +126,7 @@ namespace JSIStudios.SimpleRESTServices.Server
                 if (WebOperationContext.Current != null)
                     WebOperationContext.Current.OutgoingResponse.StatusCode = HttpStatusCode.InternalServerError;
 
-                ThrowWebFaultException<string>(string.Format("There was an error processing the request:{0}", ex.Message), HttpStatusCode.InternalServerError);
+                SetHttpErrorStatusCode(string.Format("There was an error processing the request:{0}", ex.Message), HttpStatusCode.InternalServerError);
             }
         }
 
@@ -186,115 +139,21 @@ namespace JSIStudios.SimpleRESTServices.Server
             return sb.ToString();
         }
 
-        private bool IsRequestAuthorized()
-        {
-            throw new NotImplementedException();
-        }
-
-        private void ThrowWebFaultException<T>(T value, HttpStatusCode statusCode)
+        protected virtual void SetHttpErrorStatusCode<T>(T value, HttpStatusCode statusCode)
         {
             throw new WebFaultException<T>(value, statusCode);
         }
 
-        private void ThrowWebFaultException(HttpStatusCode statusCode)
+        protected virtual void SetHttpErrorStatusCode(string content, HttpStatusCode statusCode)
+        {
+            throw new WebFaultException<string>(content, statusCode);
+        }
+
+        protected virtual void SetHttpErrorStatusCode(HttpStatusCode statusCode)
         {
             throw new WebFaultException(statusCode);
         }
 
         #endregion
-    }
-
-
-    public class RESTRequestStartedEventArgs : EventArgs
-    {
-        public Guid RequestId { get; private set; }
-
-        public string Request { get; private set; }
-
-        public RESTRequestStartedEventArgs(Guid requestId, string request)
-            : base()
-        {
-            RequestId = requestId;
-            Request = request;
-        }
-    }
-
-    public class RESTRequestErrorEventArgs : EventArgs
-    {
-        public Guid RequestId { get; private set; }
-
-        public Exception Error { get; private set; }
-
-        public RESTRequestErrorEventArgs(Guid requestId, Exception error)
-            : base()
-        {
-            RequestId = requestId;
-            Error = error;
-        }
-    }
-
-    public class RESTRequestCompletedEventArgs : EventArgs
-    {
-        public Guid RequestId { get; private set; }
-
-        public object Response { get; private set; }
-
-        public long ExecutionTime { get; private set; }
-
-        public RESTRequestCompletedEventArgs(Guid requestId, object response, long exectionTime)
-            : base()
-        {
-            RequestId = requestId;
-            Response = response;
-            ExecutionTime = exectionTime;
-
-        }
-    }
-
-    public class BadWebRequestException : Exception
-    {
-        public BadWebRequestException(string message)
-            : base(message)
-        {
-        }
-
-        public BadWebRequestException(string message, Exception innerException)
-            : base(message, innerException)
-        {
-        }
-    }
-
-    public class HttpHeaderNotFoundException : Exception
-    {
-        public string Name { get; set; }
-        public HttpHeaderNotFoundException(string name, string message)
-            : base(message)
-        {
-            Name = name;
-        }
-
-        public HttpHeaderNotFoundException(string name)
-        {
-            Name = name;
-        }
-
-        public HttpHeaderNotFoundException()
-        {
-        }
-    }
-
-    public class HttpResourceNotFoundException : Exception
-    {
-        public HttpResourceNotFoundException(string message)
-            : base(message)
-        {
-        }
-    }
-
-    public class HttpResourceNotModifiedException : Exception
-    {
-        public HttpResourceNotModifiedException()
-        {
-        }
     }
 }
